+++ evm_arithmetization/src/cpu/kernel/tests/add11.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 use std::str::FromStr;
 
+use env_logger::{try_init_from_env, Env, DEFAULT_FILTER_ENV};
 use ethereum_types::{Address, BigEndianHash, H256};
 use hex_literal::hex;
 use keccak_hash::keccak;
@@ -157,7 +158,7 @@ fn test_add11_yml() {
 
     let initial_stack = vec![];
     let mut interpreter: Interpreter<F> =
-        Interpreter::new_with_generation_inputs_and_kernel(0, initial_stack, tries_inputs);
+        Interpreter::new_with_generation_inputs(0, initial_stack, tries_inputs);
 
     let route_txn_label = KERNEL.global_labels["main"];
     // Switch context and initialize memory with the data we need for the tests.
@@ -301,7 +302,7 @@ fn test_add11_yml_with_exception() {
 
     let initial_stack = vec![];
     let mut interpreter: Interpreter<F> =
-        Interpreter::new_with_generation_inputs_and_kernel(0, initial_stack, tries_inputs);
+        Interpreter::new_with_generation_inputs(0, initial_stack, tries_inputs);
 
     let route_txn_label = KERNEL.global_labels["main"];
     // Switch context and initialize memory with the data we need for the tests.