+++ evm_arithmetization/src/cpu/kernel/tests/core/access_lists.rs
@@ -1,4 +1,4 @@
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
 
 use anyhow::Result;
 use ethereum_types::{Address, H160, U256};
@@ -21,18 +21,17 @@ fn test_init_access_lists() -> Result<()> {
 
     // Check the initial state of the access list in the kernel.
     let initial_stack = vec![0xdeadbeefu32.into()];
-    let mut interpreter = Interpreter::<F>::new_with_kernel(init_label, initial_stack);
+    let mut interpreter = Interpreter::<F>::new(init_label, initial_stack);
     interpreter.run()?;
 
     assert!(interpreter.stack().is_empty());
 
     let acc_addr_list: Vec<U256> = (0..2)
         .map(|i| {
-            interpreter.generation_state.memory.get(MemoryAddress::new(
-                0,
-                Segment::AccessedAddresses,
-                i,
-            ))
+            interpreter
+                .generation_state
+                .memory
+                .get_with_init(MemoryAddress::new(0, Segment::AccessedAddresses, i))
         })
         .collect();
     assert_eq!(
@@ -42,11 +41,10 @@ fn test_init_access_lists() -> Result<()> {
 
     let acc_storage_keys: Vec<U256> = (0..4)
         .map(|i| {
-            interpreter.generation_state.memory.get(MemoryAddress::new(
-                0,
-                Segment::AccessedStorageKeys,
-                i,
-            ))
+            interpreter
+                .generation_state
+                .memory
+                .get_with_init(MemoryAddress::new(0, Segment::AccessedStorageKeys, i))
         })
         .collect();
 
@@ -68,7 +66,7 @@ fn test_list_iterator() -> Result<()> {
     let init_label = KERNEL.global_labels["init_access_lists"];
 
     let initial_stack = vec![0xdeadbeefu32.into()];
-    let mut interpreter = Interpreter::<F>::new_with_kernel(init_label, initial_stack);
+    let mut interpreter = Interpreter::<F>::new(init_label, initial_stack);
     interpreter.run()?;
 
     // test the list iterator
@@ -95,7 +93,7 @@ fn test_insert_address() -> Result<()> {
 
     // Test for address already in list.
     let initial_stack = vec![0xdeadbeefu32.into()];
-    let mut interpreter = Interpreter::<F>::new_with_kernel(init_label, initial_stack);
+    let mut interpreter = Interpreter::<F>::new(init_label, initial_stack);
     interpreter.run()?;
 
     let insert_accessed_addresses = KERNEL.global_labels["insert_accessed_addresses"];
@@ -113,10 +111,9 @@ fn test_insert_address() -> Result<()> {
     interpreter.run()?;
     assert_eq!(interpreter.stack(), &[U256::one()]);
     assert_eq!(
-        interpreter
-            .generation_state
-            .memory
-            .get(MemoryAddress::new_bundle(U256::from(AccessedAddressesLen as usize)).unwrap()),
+        interpreter.generation_state.memory.get_with_init(
+            MemoryAddress::new_bundle(U256::from(AccessedAddressesLen as usize)).unwrap(),
+        ),
         U256::from(Segment::AccessedAddresses as usize + 4)
     );
 
@@ -129,7 +126,7 @@ fn test_insert_accessed_addresses() -> Result<()> {
 
     // Test for address already in list.
     let initial_stack = vec![0xdeadbeefu32.into()];
-    let mut interpreter = Interpreter::<F>::new_with_kernel(init_access_lists, initial_stack);
+    let mut interpreter = Interpreter::<F>::new(init_access_lists, initial_stack);
     interpreter.run()?;
 
     let insert_accessed_addresses = KERNEL.global_labels["insert_accessed_addresses"];
@@ -167,10 +164,9 @@ fn test_insert_accessed_addresses() -> Result<()> {
         interpreter.run()?;
         assert_eq!(interpreter.pop().unwrap(), U256::zero());
         assert_eq!(
-            interpreter
-                .generation_state
-                .memory
-                .get(MemoryAddress::new_bundle(U256::from(AccessedAddressesLen as usize)).unwrap()),
+            interpreter.generation_state.memory.get_with_init(
+                MemoryAddress::new_bundle(U256::from(AccessedAddressesLen as usize)).unwrap(),
+            ),
             U256::from(offset + 2 * (n + 1))
         );
     }
@@ -183,18 +179,16 @@ fn test_insert_accessed_addresses() -> Result<()> {
     interpreter.run()?;
     assert_eq!(interpreter.stack(), &[U256::one()]);
     assert_eq!(
-        interpreter
-            .generation_state
-            .memory
-            .get(MemoryAddress::new_bundle(U256::from(AccessedAddressesLen as usize)).unwrap()),
+        interpreter.generation_state.memory.get_with_init(
+            MemoryAddress::new_bundle(U256::from(AccessedAddressesLen as usize)).unwrap(),
+        ),
         U256::from(offset + 2 * (n + 2))
     );
     assert_eq!(
-        interpreter.generation_state.memory.get(MemoryAddress::new(
-            0,
-            AccessedAddresses,
-            2 * (n + 1)
-        )),
+        interpreter
+            .generation_state
+            .memory
+            .get_with_init(MemoryAddress::new(0, AccessedAddresses, 2 * (n + 1)),),
         U256::from(addr_not_in_list.0.as_slice())
     );
 
@@ -207,7 +201,7 @@ fn test_insert_accessed_storage_keys() -> Result<()> {
 
     // Test for address already in list.
     let initial_stack = vec![0xdeadbeefu32.into()];
-    let mut interpreter = Interpreter::<F>::new_with_kernel(init_access_lists, initial_stack);
+    let mut interpreter = Interpreter::<F>::new(init_access_lists, initial_stack);
     interpreter.run()?;
 
     let insert_accessed_storage_keys = KERNEL.global_labels["insert_accessed_storage_keys"];
@@ -254,8 +248,8 @@ fn test_insert_accessed_storage_keys() -> Result<()> {
         assert_eq!(interpreter.pop().unwrap(), U256::zero());
         assert_eq!(interpreter.pop().unwrap(), value);
         assert_eq!(
-            interpreter.generation_state.memory.get(
-                MemoryAddress::new_bundle(U256::from(AccessedStorageKeysLen as usize)).unwrap()
+            interpreter.generation_state.memory.get_with_init(
+                MemoryAddress::new_bundle(U256::from(AccessedStorageKeysLen as usize)).unwrap(),
             ),
             U256::from(offset + 4 * (n + 1))
         );
@@ -274,34 +268,30 @@ fn test_insert_accessed_storage_keys() -> Result<()> {
         &[storage_key_not_in_list.2, U256::one()]
     );
     assert_eq!(
-        interpreter
-            .generation_state
-            .memory
-            .get(MemoryAddress::new_bundle(U256::from(AccessedStorageKeysLen as usize)).unwrap()),
+        interpreter.generation_state.memory.get_with_init(
+            MemoryAddress::new_bundle(U256::from(AccessedStorageKeysLen as usize)).unwrap(),
+        ),
         U256::from(offset + 4 * (n + 2))
     );
     assert_eq!(
-        interpreter.generation_state.memory.get(MemoryAddress::new(
-            0,
-            AccessedStorageKeys,
-            4 * (n + 1)
-        )),
+        interpreter
+            .generation_state
+            .memory
+            .get_with_init(MemoryAddress::new(0, AccessedStorageKeys, 4 * (n + 1)),),
         U256::from(storage_key_not_in_list.0 .0.as_slice())
     );
     assert_eq!(
-        interpreter.generation_state.memory.get(MemoryAddress::new(
-            0,
-            AccessedStorageKeys,
-            4 * (n + 1) + 1
-        )),
+        interpreter
+            .generation_state
+            .memory
+            .get_with_init(MemoryAddress::new(0, AccessedStorageKeys, 4 * (n + 1) + 1),),
         storage_key_not_in_list.1
     );
     assert_eq!(
-        interpreter.generation_state.memory.get(MemoryAddress::new(
-            0,
-            AccessedStorageKeys,
-            4 * (n + 1) + 2
-        )),
+        interpreter
+            .generation_state
+            .memory
+            .get_with_init(MemoryAddress::new(0, AccessedStorageKeys, 4 * (n + 1) + 2),),
         storage_key_not_in_list.2
     );