+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 
 ### Changed
 - Add verification for invalid jumps. [#36](https://github.com/0xPolygonZero/zk_evm/pull/36)
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix running doctests in release mode ([#60](https://github.com/0xPolygonZero/zk_evm/pull/60))
 - Fix block padding without withdrawals ([#63](https://github.com/0xPolygonZero/zk_evm/pull/63))
 - Change position of empty node encoding in RLP segment ([#62](https://github.com/0xPolygonZero/zk_evm/pull/62))
+- Unify interpreter and prover witness generation ([#56](https://github.com/0xPolygonZero/zk_evm/pull/56))
 
 ## [0.1.0] - 2024-02-21
 * Initial release.